+++ test/libsolidity/syntaxTests/memberLookup/memory_structs_with_mappings.sol
@@ -7,5 +7,4 @@ contract Test {
 	}
 }
 // ----
-// TypeError: (104-114): Type struct Test.S memory is only valid in storage.
-// TypeError: (118-121): Member "b" is not available in struct Test.S memory outside of storage.
+// TypeError: (104-114): Type struct Test.S is only valid in storage.