+++ test/libsolidity/syntaxTests/types/mapping/function_type_argument_array.sol
@@ -1,0 +1,6 @@
+contract test {
+    function f(mapping(uint => uint)[2] memory b) internal {
+    }
+}
+// ----
+// TypeError: (31-64): Type mapping(uint256 => uint256)[2] is only valid in storage.