+++ test/libsolidity/syntaxTests/types/mapping/mapping_array_data_location_function_param_external.sol
@@ -2,5 +2,5 @@ contract c {
     function f1(mapping(uint => uint)[] calldata) pure external {}
 }
 // ----
-// TypeError: (29-61): Type is required to live outside storage.
-// TypeError: (29-61): Only libraries are allowed to use the mapping type in public or external functions.
+// TypeError: (29-61): Types containing (nested) mappings can only be used in storage.
+// TypeError: (29-61): Only libraries are allowed to use a (nested) mapping type in public or external functions.