+++ test/libsolidity/syntaxTests/types/mapping/mapping_struct_recursive_data_location_memory.sol
@@ -6,5 +6,5 @@ contract C {
     function f(U memory) public {}
 }
 // ----
-// TypeError: (148-156): Type is required to live outside storage.
-// TypeError: (148-156): Only libraries are allowed to use the mapping type in public or external functions.
+// TypeError: (148-156): Types containing (nested) mappings can only be used in storage.
+// TypeError: (148-156): Only libraries are allowed to use a (nested) mapping type in public or external functions.