+++ Changelog.md
@@ -5,7 +5,7 @@ Breaking changes:
  * Deprecated dot syntax for `value` and `gas`.
  * Deprecated the identifier `now`.
  * JSON AST: Removes members with ``null`` value from JSON output.
- * Type Checker: Disallows nested ``mapping`` types that are requested to live outside storage.
+ * Type Checker: Disallows structs in memory that contains a (nested) ``mapping`` type.
 
 Language Features: